Generate foreach in a way that supports queues, bug1643
The logic generated for the foreach iterations does not work with
empty queues as it will iterate from 0 to -1:

    while(($left(q)<=$right(q)) ? (i<=$right(q)) : (i>=$left(q))) {
      ...
    }

$right(q) will return -1 when q is empty.

Hence, add another check for the size and make it part of the
condition:

    while(($size(q)>0) &
          ($left(q)<=$right(q)) ? (i<=$right(q)) : (i>=$left(q))) {
       ...
    }

$size(q) will evaluate to a constant, and the part will be optimized
away. Similarly, the iteration condition could be simpler for queues,
as it can only iterate up, but at this stage we don't know the type of
the array.